requirements.txt
amqp==5.1.1
anyio==3.6.2
APScheduler==3.9.1.post1
asgiref==3.6.0
async-timeout==4.0.2
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
comments==1.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Django==3.2
django-apscheduler==0.6.2
django-auth-ldap==4.1.0
django-crontab==0.7.1
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
dnspython==2.3.0
drf-yasg==1.21.5
email-validator==2.0.0.post1
fastapi==0.95.1
fasteners==0.18
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
jenkins-job-builder==5.0.2
Jinja2==3.1.2
kombu==5.3.0
ldap3==2.9.1
MarkupSafe==2.1.2
multi-key-dict==2.0.3
numpy==1.24.2
openapi-codec==1.3.2
orjson==3.8.10
packaging==23.0
paramiko==3.1.0
pbr==5.11.1
pexpect==4.8.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jenkins==1.8.0
python-ldap==3.4.3
python-multipart==0.0.6
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.4.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.26.1
stevedore==5.0.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
Werkzeug==2.2.3
